Rename providers to services
