+++ api/src/routes/v1/proposals/list.ts
@@ -1,5 +1,5 @@
 import { OpenAPIHono, createRoute, z } from "@hono/zod-openapi";
-import { getProposals } from "@src/providers/apiNodeProvider";
+import { getProposals } from "@src/services/external/apiNodeService";
 
 const route = createRoute({
   method: "get",
