+++ api/src/routes/v1/providers/list.ts
@@ -1,6 +1,6 @@
 import { OpenAPIHono, createRoute, z } from "@hono/zod-openapi";
 import { cacheKeys, cacheResponse } from "@src/caching/helpers";
-import { getProviderList } from "@src/providers/providerStatusProvider";
+import { getProviderList } from "@src/services/db/providerStatusService";
 
 const route = createRoute({
   method: "get",
